complete example class; add tests and snippets
